fix: Rename date generator file and move it to shared folder
